+++ src/onegov.activity/onegov/activity/models/activity.py
@@ -95,14 +95,22 @@ class Activity(Base, ContentMixin, TimestampMixin):
         assert self.state == 'preview'
         self.state = 'proposed'
 
+        return self
+
     def accept(self):
         assert self.state == 'proposed'
         self.state = 'accepted'
+
+        return self
 
     def deny(self):
         assert self.state == 'proposed'
         self.state = 'denied'
 
+        return self
+
     def archive(self):
         assert self.state == 'accepted'
         self.state = 'archived'
+
+        return self