+++ src/onegov.activity/HISTORY.rst
@@ -1,6 +1,9 @@
 Changelog
 ---------
 
+- Adds the ability to filter the collection by state.
+  [href]
+
 - Adds the ability to get the set of used activity tags.
   [href]